+++ examples/ex_chroma_subsampling.py
@@ -20,9 +20,6 @@ def resize_array(img, dsize):
     #return np.array( PIL.Image.fromarray(img).resize(dsize, resample=PIL.Image.LANCZOS) )
     return cv2.resize( img, dsize=dsize, interpolation=cv2.INTER_LANCZOS4 )
 
-# For debugging only
-# from gfxdisp.pfs.pfs import pfs
-
 # This is a 16-bit image, convert to float
 I_ref_np16 = pycvvdp.load_image_as_array(os.path.join('example_media', 'wavy_facade.png'))
 
@@ -31,7 +28,7 @@ I_ref = I_ref_np16[-patch_sz:,-patch_sz:,:].astype(np.float32) / np.iinfo(I_ref_
 
 # I_ref = reshuffle_dims( torch.as_tensor(I_ref_np), "HWC", "CFHW" )
 # ct = ColorTransform()
-# I_DKL = ct.rgb2colourspace(I_ref, "DKLd65")
+# I_DKL = ct.rgb2colorspace(I_ref, "DKLd65")
 
 # sigma = 4
 # GB = tt.GaussianBlur(math.ceil(sigma*3)*2+1, sigma)
@@ -109,7 +106,7 @@ fig, axs = plt.subplots(len(ss_type), N+M, layout="constrained", figsize=(18, 8)
 if len(ss_type) == 1:
     axs = [axs]
 
-# qpp = { 'cvvdp': { 'ylabel': 'ColourVideoVDP JOD', 'ylim': (4, 10), 'col': 0 }, 
+# qpp = { 'cvvdp': { 'ylabel': 'ColorVideoVDP JOD', 'ylim': (4, 10), 'col': 0 }, 
 #         'ssim-lum': { 'ylabel': 'SSIM-luma', 'ylim': (0.4, 1), 'col': 1 },
 #         'ssim-rgb': { 'ylabel': 'SSIM-RGB', 'ylim': (0.4, 1), 'col': 1 } }